Implement debug log with Timber library
A debug logger was added to the 'MyApplication.kt' file using the Timber library. This change was made to improve debugging experience by sending logs to the Logcat in debug mode only. The 'isDebuggable' variable was added to confirm whether the app is currently in debug mode or not, planting the logger only when it's the case. Timber library was also added to the app's dependencies in 'build.gradle.kts' file.

Please remember to disable debug mode in release build type to avoid logging sensitive information.