Remove unneeded Android Log import
The import for Android's Log was removed from 'MyApplication.kt'. This was done because the app's logging has been refactored to use the Timber library ensuring cleaner, more efficient logging. A side effect of this change is also the reduction in dependencies, as we no longer need to import Android's Log class.