Add LICENSE file to the project
A LICENSE file has been added to the project to clearly define the legal terms under which the software can be used. This step has been taken to ensure that the software is free and unencumbered, and can be utilized by any individual or entity for any purpose they desire.