Update Kotlin version and add Dagger Hilt dependency
This commit updates the Kotlin version from 1.8.10 to 1.9.0 while also changing the source and target compatibility to JavaVersion 17. Along with this, the Dagger Hilt dependency has been added with version 2.47 and is applied in the top-level and app-level build.gradle.kts files. The 'kapt' has also been introduced in app-level build.gradle.kts for annotation processing. These changes are necessary for enhancing the app's performance and applying dependency injection with Dagger Hilt.