Add Screen enumeration and navigate extension
Added a Screen enumeration in 'Screen.kt' file to handle different screens, and a navigation controller extension for navigation. This changes introduces a structured way of handling screens and navigation in the app. New classes for Argument and Parameter were created to improve handling of arguments and parameters while routing in the app.