Add HomeScreen.kt for main app view
Created a new file HomeScreen.kt under the screen package for presenting the main view of the app. This is the Composable function for the home screen which displays a Column with a Text widget "Home Screen". Included a preview function for a quick visual of how it looks in the Android Studio.