Add IHomeViewModel interface for Home state management
Added a new Interface 'IHomeViewModel' under domain package of the Android application. This interface includes a 'state' val of 'HomeState' type and an 'updateState' function. It is designed to handle state changes and updates of 'HomeState' for the Home page. This will provide a better structure for state manipulation and managing data flow from UI. It also includes a static 'Preview' object for testing and previewing purposes.