Add HomeViewModel for Home state management
Added a new class 'HomeViewModel' under the implementation package. 'HomeViewModel' implements 'IHomeViewModel' and provides actual implementation for managing the state of Home page of the app. The ViewModel uses 'StateFlow' to propagate the state changes to the UI. This approach ensures that state updates are delivered to the UI components every time the state changes, improving the reactivity of the UI.