Refactor navigation from HomeScreen to DetailScreen
Removed the clickable modifier in HomeScreen which previously was responsible for navigating to DetailScreen. Instead, navigation is now handled via an onNextClick callback function. The callback function takes a message and value as parameters which are passed into the navigation method. This allows greater flexibility in navigation as the data passed can now be modified by the user.