Remove ExampleInstrumentedTest.kt
The ExampleInstrumentedTest.kt file was removed as it was used for initial testing during development stage. The test was not applicable anymore after the application structure was refined. It did not serve its purpose of asserting that the application context package name equals the given hardcoded string. More comprehensive tests covering broader functionality will be implemented as next step.