Add Back Button To DetailScreen and functionality to MainNavGraph
The DetailScreen has been updated to include a Back Button using the IconButton() component. This button allows users to navigate back from the DetailScreen. The inclusion of the back button improves UX by allowing users to easily navigate backwards in the app. The onClick() Event for the back button is handled in the MainNavGraph, where the function navController.popBackStack() is used to remove the last destination from the back stack and navigate to the destination that was on top of Stack.