Adjust app icon and launch screen settings
This update modifies the app's icon and launch screen look by changing drawable resources to color resources. The background color of the icon is now defined in a separate XML file, which provides an easier way to change the color. The size and color of the foreground icon are also adjusted to provide better aesthetics and visibility.