Add Splash Screen theme for Android 31 version.
Introduced a new theme for Android 31 version, which focuses on enhancing the application's initial splash screen. The splash screen is now based on Material Light Theme without any ActionBar. It uses a neutral background color, and the application's logo in the foreground with an animation duration of 1000ms. The logo's background color is set to the primary color of our app. This splash screen will provide a smoother visual transition during application's startup.